--- src/app/core/entity-components/historical-data/historical-entity-data.ts
@@ -1,9 +1,0 @@
-import { Entity } from "../../entity/entity";
-import { DatabaseEntity } from "../../entity/database-entity.decorator";
-import { DatabaseField } from "../../entity/database-field.decorator";
-
-@DatabaseEntity("HistoricalEntityData")
-export class HistoricalEntityData extends Entity {
-  @DatabaseField() date: Date;
-  @DatabaseField() relatedEntity: string;
-}
