+++ src/app/core/view/dynamic-components-map.ts
@@ -27,7 +27,7 @@ import { ChildrenBmiDashboardComponent } from "../../child-dev-project/children/
 import { DashboardShortcutWidgetComponent } from "../dashboard-shortcut-widget/dashboard-shortcut-widget/dashboard-shortcut-widget.component";
 import { UsersBlockComponent } from "../user/users-block/users-block.component";
 import { UserListComponent } from "../admin/user-list/user-list.component";
-import { HistoricalDataComponent } from "../entity-components/historical-data/historical-data/historical-data.component";
+import { HistoricalDataComponent } from "../../features/historical-data/historical-data/historical-data.component";
 
 export const DYNAMIC_COMPONENTS_MAP = new Map<string, any>([
   ["ChildrenCountDashboard", ChildrenCountDashboardComponent],
